requirements.txt
streamlit
openai=0.28
pillow
requests
